+++ src/om_inputs/schemas.cljs
@@ -48,7 +48,8 @@
    (s/optional-key :validations) s/Any
    (s/optional-key :action) {(s/optional-key :one-shot) s/Bool
                              (s/optional-key :no-reset) s/Bool
-                             (s/optional-key :async) s/Bool}
+                             (s/optional-key :async) s/Bool
+                             (s/optional-key :attrs) {s/Any s/Any}}
    s/Keyword {(s/optional-key :type)      s/Str
               (s/optional-key :labeled)   s/Bool
               (s/optional-key :attrs)     s/Any